feat(mrols): migrate language server functionality
This patch migrates the implementation of the Python MRO language server
from the original project, in order separate the VS Code extension
development and the language server development.

This migrated version implements the necessary foundations of the
language server, which includes the competences like replying to
initialisation and termination request, reacting to document changes,
and reply to the hover or code lens request with MRO information.