test(tests and examples): migrate tests & examples
This patch migrates the unit tests and the test examples from the
original project.

The unit tests are all in the `tests` folder and the test examples are all in the `test_data` folder.

These tests and examples cover the normal cases and some edge cases like syntax error, class declaration in comments, conflict MRO, etc.